Acquire the manager lock for all read/write operations.
This makes the standard task operations thread-safe, rather than requiring
callers to lock everything themselves. The recursive lock makes it safe to
nest functions that acquire the lock.

Note that this also fixes a race condition between a task calling
REMOVE-TASK on completion and the task being added to the manager's task
list. Previously, the task could complete before the task had been added,
which caused REMOVE-TASK to do nothing; afterwards, the now-defunct task
would be added to the list, resulting in a zombie.

The lock acquisition in REMOVE-TASK and around the addition of the task
means the task can't remove itself from the list until the acceptor task
has finished adding it any released the lock.